+++ CHANGELOG.md
@@ -5,6 +5,14 @@ View detailed changes in our [Trello board](https://trello.com/b/EPJJJbYH/avalan
 
 ## [Unreleased]
 
+## [0.9.0] - 2021-05-08
+### Added
+- Added way to load models right from WebNJS by importing 'webnjs/models'.
+- Added global project context variable `project` to runtime.
+
+### Changes
+- Fixes and improvements
+
 ## [0.8.10] - 2021-05-08
 ### Changes
 - Mainly fixes and improvements